# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "A comprehensive job portal API with RSA-512 signed JWT authentication. The system uses separate access and refresh tokens stored in a database with automatic revocation and blacklisting capabilities.",
        "contact": {
            "name": "Job Portal API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Job Portal API Server - All endpoints are prefixed with /api/v1"
        }
    ],
    "paths": {
...
        "/occupations": {
            "get": {
                "tags": [
                    "Occupations"
                ],
                "summary": "Get paginated list of occupations",
                "description": "Retrieve a paginated list of occupations with filtering, searching, and sorting capabilities. Supports full-text search by occupation name.",
                "operationId": "82221ea60dbba94666619150e93540ed",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (starts from 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Number of items per page (1-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 10
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "created_at",
                                "updated_at"
                            ],
                            "example": "id"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter occupations by name (case-insensitive partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Engineer"
                        }
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Preferred language for response messages",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of occupations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Array of occupation objects",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Occupation"
                                            }
                                        },
                                        "links": {
                                            "description": "Pagination links",
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/occupations?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/occupations?page=25"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/occupations?page=2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "description": "Pagination metadata",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 245
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Invalid query parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/occupations/{occupation}": {
            "get": {
                "tags": [
                    "Occupations"
                ],
                "summary": "Get specific occupation by ID",
                "description": "Retrieve a specific occupation by its ID. Returns basic occupation information without related industry data to keep response lightweight.",
                "operationId": "5391d690ac9e5aa27a961d826f0cdd31",
                "parameters": [
                    {
                        "name": "occupation",
                        "in": "path",
                        "description": "Occupation ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Preferred language for response messages",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Occupation retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Occupation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Occupation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating the occupation was not found",
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Occupation] 999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ValidationError": {
                "title": "Validation Error Response",
                "description": "Standard Laravel validation error response",
                "properties": {
                    "message": {
                        "description": "Human readable error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Field-specific validation errors",
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required.",
                                "The email must be a valid email address."
                            ],
                            "password": [
                                "The password must be at least 8 characters."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "Industry": {
                "title": "Industry",
                "description": "Industry model representing business sectors",
                "properties": {
                    "id": {
                        "description": "Unique industry identifier",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Industry name",
                        "type": "string",
                        "example": "Technology"
                    },
                    "created_at": {
                        "description": "Timestamp when the industry was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the industry was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "occupations": {
                        "description": "Array of related occupations (only included when relations[occupations] is requested or in show endpoint)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Occupation"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Occupation": {
                "title": "Occupation",
                "description": "Occupation model representing job roles within industries",
                "properties": {
                    "id": {
                        "description": "Unique occupation identifier",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Occupation name",
                        "type": "string",
                        "example": "Software Engineer"
                    },
                    "industry_id": {
                        "description": "ID of the industry this occupation belongs to",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Timestamp when the occupation was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the occupation was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00.000000Z"
                    },
                    "industry": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Industry"
                            }
                        ],
                        "nullable": true,
                        "description": "Related industry object (when relationship is loaded)"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token"
            },
            "refreshAuth": {
                "type": "http",
                "description": "JWT Refresh Token - Use the refresh token to get new access tokens. Format: Bearer {refresh_jwt_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}